requirements.txt
cffi==1.6.0
cryptography==1.3.1
enum34==1.1.4
github3.py==0.9.5
idna==2.1
ipaddress==1.0.16
jira==1.0.3
ndg-httpsclient==0.4.0
oauthlib==1.1.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
requests==2.10.0
requests-oauthlib==0.6.1
requests-toolbelt==0.6.0
six==1.10.0
tlslite==0.4.9
uritemplate.py==0.3.0
uWSGI==2.0.12
Werkzeug==0.11.9
